improve personal model and resolve the problem of switch

+++ android/kreader/app/oem/JDRead/src/main/java/com/onyx/jdread/personal/cloud/entity/jdbean/GetRechargePackageBean.java
@@ -13,7 +13,7 @@ public class GetRechargePackageBean {
 
     public static class DataBean {
         public int package_id;
-        public int recharge_money;
+        public double recharge_money;
         public int voucher_amount;
     }
 }
